+++ uni_modules/uview-ultra/libs/i18n/locales/zh-Hans.js
@@ -1,4 +1,4 @@
-{
+export default {
     "up_common_cancel": "取消",
     "up_common_confirm": "确定",
     "up_common_start": "开始",